moves tests to their own folders, completes the main logic for sign up
