+++ lib/ts/recipe/emailpassword/api/signup.ts
@@ -17,6 +17,8 @@ import Recipe from "../recipe";
 import { Request, Response, NextFunction } from "express";
 import { normaliseEmail } from "../utils";
 import { FORM_FIELD_EMAIL_ID, FORM_FIELD_PASSWORD_ID } from "../constants";
+import Session from "../../session";
+import { send200Response } from "../../../utils";
 
 export async function signUpAPI(recipeInstance: Recipe, req: Request, res: Response, next: NextFunction) {
     // Logic as per https://github.com/supertokens/supertokens-node/issues/21#issuecomment-710423536
@@ -41,7 +43,11 @@ export async function signUpAPI(recipeInstance: Recipe, req: Request, res: Respo
     );
 
     // step 5
-    // TODO: Create a new session using the Session recipe and return OK
+    await Session.createNewSession(res, user.id);
+    return send200Response(res, {
+        status: "OK",
+        user,
+    });
 }
 
 async function validateFormFieldsOrThrowError(
